Add remote state merging/saving via wsgi server
Adds a wsgi server for remote saving and merging. This allows multiple
clients to be used, as well as more persistent state. This drastically
changes how user settings is saved, as the active client is no longer
the only source for state changes.

This also adds s3-type remote storage for cloud availability and breaks
up the monolith that was script.js.